+++ com.peopleground.sagwim.app/app/(app)/profile-edit-address.tsx
@@ -1,36 +1,22 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import React, { useCallback, useMemo, useState } from 'react'
 import {
   ActivityIndicator,
   Alert,
-  FlatList,
   Pressable,
   StyleSheet,
   Text,
-  TextInput,
   View,
 } from 'react-native'
 import { Stack, router, useLocalSearchParams } from 'expo-router'
-import { Ionicons } from '@expo/vector-icons'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
 import { useAuth } from '../../src/context/AuthContext'
 import { updateMyProfile } from '../../src/api/userApi'
-import apiClient from '../../src/lib/apiClient'
 import { ConfirmDialog } from '../../src/components/common/ConfirmDialog'
+import { AddressSearchInput } from '../../src/components/common/AddressSearchInput'
 import { fontSize, radius, spacing } from '../../src/constants/theme'
 import { useTheme } from '../../src/context/ThemeContext'
 import type { UserDetailResponse } from '../../src/types/user'
-
-async function searchAddress(query: string): Promise<string[]> {
-  try {
-    const response = await apiClient.get<{ suggestions: string[] }>('/address/search', {
-      params: { query },
-    })
-    return response.data.suggestions ?? []
-  } catch {
-    return []
-  }
-}
 
 export default function ProfileEditAddressScreen() {
   const insets = useSafeAreaInsets()
@@ -39,14 +25,9 @@ export default function ProfileEditAddressScreen() {
   const { profile: profileJson } = useLocalSearchParams<{ profile: string }>()
   const profile: UserDetailResponse = JSON.parse(profileJson ?? '{}')
 
-  const [query, setQuery] = useState('')
-  const [suggestions, setSuggestions] = useState<string[]>([])
-  const [isSearching, setIsSearching] = useState(false)
   const [selectedAddress, setSelectedAddress] = useState(profile.address ?? '')
   const [saving, setSaving] = useState(false)
   const [showDiscard, setShowDiscard] = useState(false)
-
-  const debounceTimer = useRef<ReturnType<typeof setTimeout> | null>(null)
 
   const isChanged = selectedAddress.trim() !== (profile.address ?? '').trim()
   const canSave = isChanged && selectedAddress.trim().length > 0
@@ -58,31 +39,6 @@ export default function ProfileEditAddressScreen() {
     }
     router.back()
   }, [isChanged])
-
-  // 검색어 변경 시 debounce 검색
-  useEffect(() => {
-    if (debounceTimer.current) clearTimeout(debounceTimer.current)
-    const trimmed = query.trim()
-    if (trimmed.length < 2) {
-      setSuggestions([])
-      return
-    }
-    debounceTimer.current = setTimeout(async () => {
-      setIsSearching(true)
-      const results = await searchAddress(trimmed)
-      setSuggestions(results)
-      setIsSearching(false)
-    }, 300)
-    return () => {
-      if (debounceTimer.current) clearTimeout(debounceTimer.current)
-    }
-  }, [query])
-
-  const handleSelectSuggestion = useCallback((addr: string) => {
-    setSelectedAddress(addr)
-    setQuery('')
-    setSuggestions([])
-  }, [])
 
   const handleSave = useCallback(async () => {
     if (!canSave) return
@@ -124,64 +80,6 @@ export default function ProfileEditAddressScreen() {
     headerBack: { fontSize: fontSize.sm, color: colors.text },
     headerTitle: { fontSize: fontSize.md, fontWeight: '600', color: colors.text },
     body: { padding: spacing.sp4, gap: spacing.sp3 },
-    selectedRow: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      gap: spacing.sp2,
-      paddingVertical: spacing.sp3,
-      paddingHorizontal: spacing.sp3,
-      borderRadius: radius.md,
-      backgroundColor: colors.surface2,
-    },
-    selectedText: {
-      flex: 1,
-      fontSize: fontSize.base,
-      color: colors.text,
-      fontWeight: '500',
-    },
-    label: { fontSize: fontSize.sm, fontWeight: '600', color: colors.text },
-    searchRow: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      position: 'relative',
-    },
-    input: {
-      flex: 1,
-      height: 44,
-      borderWidth: 1,
-      borderColor: colors.border,
-      borderRadius: radius.md,
-      paddingHorizontal: spacing.sp3,
-      paddingRight: spacing.sp8,
-      fontSize: fontSize.base,
-      color: colors.text,
-      backgroundColor: colors.bg,
-    },
-    searchIndicator: {
-      position: 'absolute',
-      right: spacing.sp3,
-    },
-    hint: { fontSize: fontSize.sm, color: colors.textMuted },
-    suggestionList: {
-      borderWidth: 1,
-      borderColor: colors.border,
-      borderRadius: radius.md,
-      overflow: 'hidden',
-      backgroundColor: colors.bg,
-    },
-    suggestionRow: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      justifyContent: 'space-between',
-      paddingHorizontal: spacing.sp4,
-      paddingVertical: spacing.sp3,
-      borderBottomWidth: 1,
-      borderBottomColor: colors.border,
-    },
-    rowPressed: { backgroundColor: colors.surface2 },
-    suggestionRowSelected: {},
-    suggestionText: { fontSize: fontSize.base, color: colors.text, flex: 1 },
-    suggestionTextSelected: { color: colors.accent, fontWeight: '600' },
     saveBtn: {
       marginTop: spacing.sp2,
       height: 48,
@@ -214,73 +112,12 @@ export default function ProfileEditAddressScreen() {
         </View>
 
         <View style={styles.body}>
-          {/* 현재 선택된 주소 */}
-          {selectedAddress.trim().length > 0 && (
-            <View style={styles.selectedRow}>
-              <Ionicons name="location" size={16} color={colors.accent} />
-              <Text style={styles.selectedText} numberOfLines={1}>
-                {selectedAddress}
-              </Text>
-            </View>
-          )}
-
-          {/* 검색 입력 */}
-          <Text style={styles.label}>주소 검색</Text>
-          <View style={styles.searchRow}>
-            <TextInput
-              style={styles.input}
-              value={query}
-              onChangeText={setQuery}
-              placeholder="동명(읍, 면)으로 검색 (예: 서초동)"
-              placeholderTextColor={colors.textMuted}
-              editable={!saving}
-              autoFocus
-              returnKeyType="search"
-            />
-            {isSearching && (
-              <ActivityIndicator
-                size="small"
-                color={colors.accent}
-                style={styles.searchIndicator}
-              />
-            )}
-          </View>
-
-          <Text style={styles.hint}>동·읍·면 단위까지만 입력해 주세요.</Text>
-          <Text style={styles.hint}>주소는 모임 추천 등에 활용됩니다.</Text>
-
-          {/* 검색 결과 목록 */}
-          {suggestions.length > 0 && (
-            <View style={styles.suggestionList}>
-              <FlatList
-                data={suggestions}
-                keyExtractor={(item) => item}
-                scrollEnabled={false}
-                renderItem={({ item }) => (
-                  <Pressable
-                    style={({ pressed }) => [
-                      styles.suggestionRow,
-                      pressed && styles.rowPressed,
-                      item === selectedAddress && styles.suggestionRowSelected,
-                    ]}
-                    onPress={() => handleSelectSuggestion(item)}
-                  >
-                    <Text
-                      style={[
-                        styles.suggestionText,
-                        item === selectedAddress && styles.suggestionTextSelected,
-                      ]}
-                    >
-                      {item}
-                    </Text>
-                    {item === selectedAddress && (
-                      <Ionicons name="checkmark" size={16} color={colors.accent} />
-                    )}
-                  </Pressable>
-                )}
-              />
-            </View>
-          )}
+          <AddressSearchInput
+            value={selectedAddress}
+            onChange={setSelectedAddress}
+            disabled={saving}
+            autoFocus
+          />
 
           {/* 저장 버튼 */}
           <Pressable
